Ensure Connection.run_behaviors() doesn't leak pending asyncio tasks
If one of our behaviors raised an exception in their post_apply(),
run_behaviors() would leak pending asyncio tasks, causing asyncio
to complain. This fixes that by ensuring all tasks are done/cancelled
