+++ mypy.ini
@@ -1,5 +1,5 @@
 [mypy]
-warn_unused_ignores = True
+warn_unused_ignores = False
 ignore_missing_imports = True
 strict_optional = False
 check_untyped_defs = True
