+++ p2p/connection.py
@@ -16,6 +16,7 @@ from typing import (
 )
 
 from async_service import Service
+from async_service.asyncio import cleanup_tasks
 
 from cached_property import cached_property
 
@@ -48,7 +49,7 @@ from p2p.exceptions import (
     UnknownProtocol,
     UnknownProtocolCommand,
 )
-from p2p.logic import wait_first
+from p2p.asyncio_utils import create_task, wait_first
 from p2p.subscription import Subscription
 from p2p.p2p_proto import BaseP2PProtocol, DevP2PReceipt, Disconnect
 from p2p.typing import Capabilities
@@ -115,22 +116,26 @@ class Connection(ConnectionAPI, Service):
     async def run_behaviors(self, behaviors: Tuple[BehaviorAPI, ...]) -> None:
         async with contextlib.AsyncExitStack() as stack:
             futures: List[asyncio.Future[None]] = [
-                asyncio.create_task(self.manager.wait_finished())]
+                create_task(self.manager.wait_finished(), 'Connection/run_behaviors/wait_finished')]
             for behavior in behaviors:
                 if behavior.should_apply_to(self):
                     behavior_exit = await stack.enter_async_context(behavior.apply(self))
                     futures.append(behavior_exit)
 
             self.behaviors_applied.set()
-            try:
-                for behavior in behaviors:
-                    behavior.post_apply()
-                await wait_first(futures)
-            except PeerConnectionLost:
-                # Any of our behaviors may propagate a PeerConnectionLost, which is to be expected
-                # as many Connection APIs used by them can raise that. To avoid a DaemonTaskExit
-                # since we're returning silently, ensure we're cancelled.
-                self.manager.cancel()
+            # If wait_first() is called, cleanup_tasks() will be a no-op, but if any post_apply()
+            # calls raise an exception, it will ensure we don't leak pending tasks that would
+            # cause asyncio to complain.
+            async with cleanup_tasks(*futures):
+                try:
+                    for behavior in behaviors:
+                        behavior.post_apply()
+                    await wait_first(futures)
+                except PeerConnectionLost:
+                    # Any of our behaviors may propagate a PeerConnectionLost, which is to be
+                    # expected as many Connection APIs used by them can raise that. To avoid a
+                    # DaemonTaskExit since we're returning silently, ensure we're cancelled.
+                    self.manager.cancel()
 
     async def run_peer(self, peer: 'BasePeer') -> None:
         """
